Create login and tweet list page

+++ app/src/main/java/com/konifar/twipe/domain/repository/TweetRepository.java
@@ -5,9 +5,9 @@ import com.konifar.twipe.domain.model.TweetModel;
 import java.util.Collection;
 
 public interface TweetRepository {
-  void getHomeTweetList(int lastTweetId, TweetListCallback callback);
+  void getHomeTweetList(Long lastTweetId, TweetListCallback callback);
 
-  void getUserTweetList(int userId, int lastTweetId, TweetListCallback callback);
+  void getUserTweetList(long userId, Long lastTweetId, TweetListCallback callback);
 
   interface TweetListCallback {
     void onTweetListLoaded(Collection<TweetModel> tweets);
